@@ -31,79 +37,48 @@
     "img",
     "image"
   ],
-  "author": "mirari",
   "license": "MIT",
+  "author": "mirari",
   "bugs": {
     "url": "https://github.com/mirari/v-viewer/issues"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mirari/v-viewer.git"
+  },
   "homepage": "https://github.com/mirari/v-viewer#readme",
-  "engines": {
-    "node": ">=4",
-    "npm": ">=3"
+  "dependencies": {
+    "lodash": "^4.17.21",
+    "viewerjs": "^1.9.0"
+  },
+  "peerDependencies": {
+    "vue": "^2.6.14"
   },
   "devDependencies": {
+    "@antfu/eslint-config": "^0.6.4",
     "@commitlint/cli": "^12.1.1",
     "@commitlint/config-conventional": "^12.1.1",
-    "autoprefixer": "^6.7.7",
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^7.1.1",
-    "babel-helper-vue-jsx-merge-props": "^2.0.2",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-component": "^0.9.1",
-    "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-export-extensions": "^6.8.0",
-    "babel-plugin-transform-vue-jsx": "^3.4.2",
-    "babel-polyfill": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-2": "^6.22.0",
-    "bulma": "^0.5.1",
-    "connect-history-api-fallback": "^1.3.0",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.27.1",
-    "electron-devtools-installer": "^2.0.1",
-    "eslint": "^4.18.2",
-    "eslint-config-standard": "^7.0.1",
-    "eslint-friendly-formatter": "^2.0.7",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-html": "^3.2.2",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^2.1.1",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.15.2",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
-    "file-loader": "^0.10.1",
-    "gh-pages": "^1.1.0",
-    "github-markdown-css": "^2.8.0",
-    "html-webpack-plugin": "^2.25.0",
-    "http-proxy-middleware": "^0.17.4",
+    "@types/lodash-es": "^4.17.4",
+    "@typescript-eslint/eslint-plugin": "^4.22.1",
+    "@typescript-eslint/parser": "^4.22.0",
+    "@vue/compiler-sfc": "^3.2.26",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "@vue/runtime-dom": "^3.2.4",
+    "bulma": "^0.9.2",
+    "eslint": "^7.25.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-vue": "^7.9.0",
     "husky": "^6.0.0",
-    "imports-loader": "^0.7.0",
     "lint-staged": "^11.0.0",
-    "node-sass": "^4.13.1",
-    "opn": "^4.0.2",
-    "ora": "^1.1.0",
-    "postcss-loader": "^1.3.3",
-    "progress-bar-webpack-plugin": "^1.9.1",
-    "sass-loader": "^6.0.3",
-    "serve-favicon": "^2.4.1",
-    "shelljs": "^0.8.3",
+    "prettier": "^2.2.1",
+    "sass": "^1.47.0",
     "standard-version": "^9.3.0",
-    "style-loader": "^0.13.1",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^3.0.1",
-    "url-loader": "^0.5.7",
-    "vue": "^2.5.6",
-    "vue-html-loader": "^1.2.3",
-    "vue-loader": "^11.1.4",
-    "vue-markdown-loader": "2.1.0",
-    "vue-template-compiler": "^2.5.6",
-    "webpack": "^2.2.1",
-    "webpack-dev-middleware": "^1.9.0",
-    "webpack-hot-middleware": "^2.14.0",
-    "webpack-merge": "^4.0.0"
-  },
-  "dependencies": {
-    "throttle-debounce": "^2.0.1",
-    "viewerjs": "^1.5.0"
+    "typescript": "^4.5.4",
+    "vite": "^2.7.10",
+    "vite-plugin-vue2": "^1.9.2",
+    "vue-template-compiler": "^2.6.14",
+    "vue-tsc": "^0.30.2"
   }
 }
