@@ -1,38 +1,52 @@
 <template>
   <div>
-    <h1 class="title">Demo</h1>
+    <h1 class="title">
+      Demo
+    </h1>
     <hr>
     <div style="margin-bottom: 3em">
-      <h2 class="title is-4">Directive</h2>
-      <p class="subtitle"><a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/DirectiveExample.vue" target="_blank">Source</a></p>
+      <h2 class="title is-4">
+        Directive
+      </h2>
+      <p class="subtitle">
+        <a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/DirectiveExample.vue" target="_blank">Source</a>
+      </p>
       <directive-example />
     </div>
     <div style="margin-bottom: 6em">
-      <h2 class="title is-4">Component</h2>
-      <p class="subtitle"><a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/ComponentExample.vue" target="_blank">Source</a></p>
+      <h2 class="title is-4">
+        Component
+      </h2>
+      <p class="subtitle">
+        <a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/ComponentExample.vue" target="_blank">Source</a>
+      </p>
       <component-example />
     </div>
     <div>
-      <h2 class="title is-4">Api</h2>
-      <p class="subtitle"><a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/ApiExample.vue" target="_blank">Source</a></p>
+      <h2 class="title is-4">
+        Api
+      </h2>
+      <p class="subtitle">
+        <a href="https://github.com/mirari/v-viewer/blob/master/example/views/example/ApiExample.vue" target="_blank">Source</a>
+      </p>
       <api-example />
     </div>
   </div>
 </template>
 
 <script>
-import DirectiveExample from './DirectiveExample'
-import ComponentExample from './ComponentExample'
-import ApiExample from './ApiExample'
+import DirectiveExample from './DirectiveExample.vue'
+import ComponentExample from './ComponentExample.vue'
+import ApiExample from './ApiExample.vue'
 
 export default {
   components: {
     ComponentExample,
     DirectiveExample,
-    ApiExample
+    ApiExample,
   },
 
-  data () {
+  data() {
     return {
     }
   },
@@ -41,9 +55,9 @@ export default {
 
   },
 
-  mounted () {
+  mounted() {
 
-  }
+  },
 }
 </script>
 
