@@ -1,9 +1,9 @@
-import Viewer from 'viewerjs'
-import { debounce } from 'throttle-debounce'
 import Vue from 'vue'
+import Viewer from 'viewerjs'
+import debounce from 'lodash/debounce'
 
-const directive = ({name = 'viewer', debug = false} = {}) => {
-  function createViewer (el, options, rebuild = false, observer = false) {
+const directive = ({ name = 'viewer', debug = false } = {}) => {
+  function createViewer(el, options, rebuild = false, observer = false) {
     // nextTick执行，否则可能漏掉未渲染完的子元素
     Vue.nextTick(() => {
       // 如果无图片或者和上次比较没有变化，那么就没有必要重新初始化或更新
@@ -12,36 +12,38 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
         destroyViewer(el)
         el[`$${name}`] = new Viewer(el, options)
         log('Viewer created')
-      } else {
+      }
+      else {
         el[`$${name}`].update()
         log('Viewer updated')
       }
     })
   }
 
-  function imageDiff (el) {
+  function imageDiff(el) {
     const imageContent = el.innerHTML.match(/<img([\w\W]+?)[\\/]?>/g)
     const viewerImageText = imageContent ? imageContent.join('') : undefined
     if (el.__viewerImageDiffCache === viewerImageText) {
       log('Element change detected, but image(s) has not changed')
       return false
-    } else {
+    }
+    else {
       log('Image change detected')
       el.__viewerImageDiffCache = viewerImageText
       return true
     }
   }
 
-  function createObserver (el, options, debouncedCreateViewer, rebuild) {
+  function createObserver(el, options, debouncedCreateViewer, rebuild) {
     destroyObserver(el)
-    const MutationObserver = global.MutationObserver || global.WebKitMutationObserver || global.MozMutationObserver
+    const MutationObserver = window.MutationObserver || window.WebKitMutationObserver || window.MozMutationObserver
     if (!MutationObserver) {
       log('Observer not supported')
       return
     }
-    const observer = new MutationObserver(function (mutations) {
-      mutations.forEach(function (mutation) {
-        log('Viewer mutation:' + mutation.type)
+    const observer = new MutationObserver((mutations) => {
+      mutations.forEach((mutation) => {
+        log(`Viewer mutation:${mutation.type}`)
         debouncedCreateViewer(el, options, rebuild, true)
       })
     })
@@ -51,22 +53,22 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
     log('Observer created')
   }
 
-  function createWatcher (el, {expression}, vnode, debouncedCreateViewer) {
+  function createWatcher(el, { expression }, vnode, debouncedCreateViewer) {
     const simplePathRE = /^[A-Za-z_$][\w$]*(?:\.[A-Za-z_$][\w$]*|\['[^']*?']|\["[^"]*?"]|\[\d+]|\[[A-Za-z_$][\w$]*])*$/
     if (!expression || !simplePathRE.test(expression)) {
       log('Only simple dot-delimited paths can create watcher')
       return
     }
-    el.__viewerUnwatch = vnode.context.$watch(expression, function (newVal, oldVal) {
+    el.__viewerUnwatch = vnode.context.$watch(expression, (newVal, oldVal) => {
       log('Change detected by watcher: ', expression)
       debouncedCreateViewer(el, newVal, true)
     }, {
-      deep: true
+      deep: true,
     })
     log('Watcher created, expression: ', expression)
   }
 
-  function destroyViewer (el) {
+  function destroyViewer(el) {
     if (!el[`$${name}`]) {
       return
     }
@@ -75,7 +77,7 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
     log('Viewer destroyed')
   }
 
-  function destroyObserver (el) {
+  function destroyObserver(el) {
     if (!el.__viewerMutationObserver) {
       return
     }
@@ -84,7 +86,7 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
     log('Observer destroyed')
   }
 
-  function destroyWatcher (el) {
+  function destroyWatcher(el) {
     if (!el.__viewerUnwatch) {
       return
     }
@@ -93,14 +95,14 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
     log('Watcher destroyed')
   }
 
-  function log () {
+  function log() {
     debug && console.log(...arguments)
   }
 
   const directive = {
-    bind (el, binding, vnode) {
+    bind(el, binding, vnode) {
       log('Viewer bind')
-      const debouncedCreateViewer = debounce(50, createViewer)
+      const debouncedCreateViewer = debounce(createViewer, 50)
       debouncedCreateViewer(el, binding.value)
 
       // 创建watch监听options表达式变化
@@ -112,7 +114,7 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
         createObserver(el, binding.value, debouncedCreateViewer, binding.modifiers.rebuild)
       }
     },
-    unbind (el, binding) {
+    unbind(el, binding) {
       log('Viewer unbind')
       // 销毁dom变化监听
       destroyObserver(el)
@@ -120,7 +122,7 @@ const directive = ({name = 'viewer', debug = false} = {}) => {
       destroyWatcher(el)
       // 销毁viewer
       destroyViewer(el)
-    }
+    },
   }
 
   return directive
